Split interface.cpp config/calib/health screens into a sibling file
interface.cpp had grown to 1265 lines mixing the high-traffic main UI
rendering with eight config/calibration/health/reset screens that
share a different vocabulary (begin/select/print menu helpers). Move
the menu-style screens into interface_config_screens.cpp and expose
the shared helpers through interface_menu_helpers.h. interface.cpp
drops to 776 lines and the new file lands at 492, both well below
the size-outlier threshold. Main UI and external display rendering
stay in interface.cpp for this round; further splits are follow-up
work, not gating.
